requirements.txt
Flask
requests
google-generativeai
gspread
oauth2client
gunicorn